config: the remaining validators move out of the record
Fourteen checks that decide nothing and only read a resolving view to raise,
placed with the family each belongs to: `lora_hook.py` (new) takes the LoRA
argument checks and the LoRA/speculative compatibility rule; `kv_cache_hook`,
`mamba_hook`, `cuda_graph_hook` and `model_path_hook` take theirs; and
`validation_hook.py` (new) takes the ones that span no single family --
`check_server_args` itself, the bucket-rule parser, the load-publish arguments,
the IB device list, the experimental marlin gate, the prefill/decode interval
and the two-batch-overlap rule.

`check_server_args` keeps its name on the record: it is a stage the launch path
calls, like `resolve_once`, so the slot is the stage and the body is what moved.

The slot forwards the return value now. `_validate_ib_devices` and
`validate_transfer_engine` both return one, and a slot that dropped it would
have handed every caller `None` -- a silent wrong answer rather than a crash.

`is_cuda` stopped being imported by the record once the last of these left, so
two tests that patched `sglang.srt.server_args.is_cuda` now patch the module
the read went to.

`server_args.py` is 6020 lines, from 6723. 62 launch shapes and the 24-shape
resolution dump are identical.
